Updated datatype on method call
git-svn-id: svn://svn.code.sf.net/p/mtsieve/svn@121 bb85bfea-1d01-4b18-b6d8-510d50b5ba24

+++ core/MpArithVector.h
@@ -168,7 +168,7 @@ public:
 	}
    
 	// Convert n to Montgomery representation
-	MpResVector<N> nToRes(uint32_t n) const
+	MpResVector<N> nToRes(uint64_t n) const
 	{
 		// n * (2^64)^2 = (n * 2^64) * (1 * 2^64)
 		MpResVector<N> r;
